Parameterizes the hard-coded Chef role names for Keystone and nova-setup
There is a lookup-by-role mechanism in use for finding node
attribute information for the node that runs the Keystone Service API
as well as the node that initially executes the Nova setup code
-- like creating the Nova database, etc. These role names are
currently hard-coded in the recipes. This patch makes that
a configurable node attribute for deployments that name their
Chef roles differently.