Patched novnc code to handle port 443 and 80
We were able to run novnc if it were on a port other than 443/80,
otherwise it would hang with "loading" text.  This bug was corrected
upstream, a few days ago.  We are likely not to see this fix get
into opentstack repos for quite some time (if ever in folsom).

https://github.com/kanaka/noVNC/pull/245