Renamed keystone_registration recipe
Renamed keystone_registration to identity_registration, and added
identity_registration tests.  Added registration tests to simply
test the args being passed to the `openstack_identity_register` LWRP.
The identity LWRP should be tested at a deeper level than the consumers.

Testing args passed to a DSL (&block) is a bit tricky. The tests
are not ideal, but give us the necessary coverage to catch common
errors.

Change-Id: I18db535ca1bab357ba083d57a76d86c5ccfddf15
Fixes: Bug 1182576