Using ~> versioning
Since we should be performing semver across our cookbooks, we should
lock to the patch version.  It means "equal to or greater than in the
last digit", so e.g. "~> 2.3" means "equal to 2.3 or greater than 2.3,
but less than 3.0", while "~> 2.3.0" would mean "equal to 2.3.0 or greater
than 2.3.0, but less than 2.4.0".
  http://stackoverflow.com/questions/5170547/what-does-tilde-greater-than-mean-in-ruby-gem-dependencies

Change-Id: I53b334865909868dcc9b5d003795adead1ecf86c