Add tailor to the compute cookbook
Add a tailor style guide check to be incorporated with the standard
run_tests. These checks are the defaults with the exception of the 80
character line limit. As this is being added to the standard set of
tests being executed during CI, this will add an extra step - locally
this was costing me ~20sec per run.

Change-Id: Ie78bc7c969e0c8d3563d76829b25c19b66341292