Update vncproxy service resource to use service name
Instead of calling the nova-vncproxy service by the generic
"nova-vncproxy" name, name this resource by its actual service name.
This change aids recieps that are dependent upon knowing which
service to notify when necessary. While it is possible to simply
notify "nova-vncproxy," there are times when that is a less favorable
solution.

Change-Id: Iefbec31b050dd93eba118012b07745dcc215f8c5