Replace config_by_role with search_for
config_by_role is almost entirely unused, and returns inconsisent
results depending on which node it is called from. Similarly, with the
new structure of these cookbooks, the "section" attribute of
config_by_role is pretty useless. So get rid of config_by_role and
replace it with search_for in this case.

Also ajust test to account for change in return type.

This change is related to https://review.openstack.org/38738

Change-Id: I12e99e665a037791ef1f8c351f261cd34ae38d72