Bug fix add_floaters subcommand processing errors
* AttributeError for args.interface Certain argument
parameters are not set depending on whether you are entering
via the neutron or nova subcommands. If we simply set
self._args, we can access the expected arguments within
the appropriate subcommand and not try to blindly
set several of them on FloatingAddress class init.

* Subcommands do not show up as arguments, so we need to
set the subcommand name to a subparser_name argument
we can check to see which subcommand was called

* If the quantum api is down for any reason, the add
add floaters script will fail raising a chef error.  We
need only raise an error if there was an error adding
the IPs, not querying the quantum api.

Change-Id: Iac2cdae03bc3e2af39781994707b5139c401427f