Bugfix public network creation
Altered the command to check for the public network to guard
against a failed command due to network errors or timeouts. Now
we will only examine the list of networks returned if we successfully
retrieved a list of networks to begin with.  We don't worry about
false positives because this is normally run when an installation is
first instantiated so there should be few if any other networks.

Change-Id: I55cc3e569afebcf6b06a5cd0d3c962da35474579