not install DHCP agent on compute node by default
From the quantum document, when there are 3 types
nodes: controller node, network node, and compute
node, it's always supposed that the network node
has OpenStack Networking L2 agent, DHCP agent and
L3 agent, and the compute node has OpenStack Networking
L2 agent and the OpenStack Compute services. There
is not the DHCP agent on the compute node.But the
current recipes always installs the DHCP agent on
the compute node by default. So remove the dhcp_agent
from the network plugins list, and the network_spec.rb
also is modified.

Closes-Bug: #1233009
Change-Id: I936a55ab055ee16b203df8479f18f3ea46a9870e