Define cephx libvirt secret for rbd volumes
This commit addresses the fact that any compute-worker needs to be

- prepared to talk to the ceph cluster (ceph.conf and ceph-common)
- a secret has to be defined.

The secret is identified by the UUID (known to cinder), as defined by a
databag item (of the "secrets" data bag) with a configurable name.

The actual key is taken from another databag item.  The username is
configured via the `rbd_user`.

The chefspec test on the proper content of the generated temporary xml
file is commented out since it won't work this way: the chefspec chef
run will have deleted the (in-memory representation of the) file when it
validates the spec, hence it's empty and the test fails.  (Not sure how
to fix that test.)

Implements: blueprint rbd-for-block-storage

Change-Id: I9eecc622b4d00c65fecfde0626f574be2b9ee934