Add rbd_* settings to nova.conf when images_type=rbd
This change adds rbd_user and rbd_secret_uuid to nova.conf when
images_type is set to rbd.  This allows for using the credentials set up
for cinder/volumes with images (ephemeral disks, not glance!) in nova.

While it's not nice to re-use cinder's cephx credentials, it seems to be
the only option when trying to use rbd for disk images and volumes,
anyway: https://bugs.launchpad.net/nova/+bug/1295601

Please note that this triggers a bug that hasn't been fixed in the
ubuntu packages: https://bugs.launchpad.net/nova/+bug/1233188 (while
having been fixed upstream).  I'm not sure how to proceed with this
issue.

Change-Id: I2fe12cee88510e1e8de223e18561d3fe87a32457
Closes-Bug: #1295177
Implements: blueprint rbd-for-block-storage