Use openstack-common VNC endpoint information
The port and interface binding attributes in the Compute
cookbook are now redundant since there is the exact same
VNC endpoint information in the common attributes now.

This patch removes the redundant attributes from Compute and
updates the places they were used. This is a small change
to the default behavior of binding to the endpoint IPs
instead of the default bind_interface (was 'lo') but
makes the behavior consistent with other endpoint binding.

Closes-Bug: 1295388
Related blueprint increase-ip-binding-flexibility

Change-Id: I1eb12aeee5459912cb5d376decb397a9efef72e4