Converted VNC service IP binding to endpoint
This change converts calling into address_for to resolve the IP to
bind the VNC server to into a call into an endpoint.  This allows the
user to either specify an interface or an IP to bind to using the
standard endpoint library.

This change is dependent on change: https://review.openstack.org/83440

blueprint increase-ip-binding-flexibility

Change-Id: I956695c473e2cd7ada8da3d4249204b86e148b71