Run nova-manage as correct user/group
We need to explicitly run nova-manage with nova user/group.
Running as root may cause files in /var/log/nova/* to be owned by
root, and this will cause nova services from being able to start (due
to being unable to write to files in that directory).

This is a non-issue when you install from packages as the package
post-install typically does a nova-manage with the correct user, and
this ensures logs are created w/ correct ownership.

Change-Id: Ibe367d95aec5f7b0244812ae67873b5370fc3ae4