Updated to start metadata-api
This starts the nova-metadata-api service on the compute
nodes when nova-networking is leveraged. This it seems that
in the compute node with nova-networking the service doesn't
start, this patch starts it up and allows for the local machines
to access the metadata-api.

This also updates the test so it confirms that the correct
process starts.

Change-Id: I7e9ce935b2d1e7135809b36684516d146277f8dc