Only have one path of support for api-metadata
The current cookbook allows the enabled_apis nova.conf key
to include the metadata value.  This causes the nove compute
service to try to auto start the api-metadata service.  But we
also have the api-metadata recipe in the run list by default which
also tries to start the api-metadata service. This can cause timing
issues with the starting service.

This patch will remove the metadata value from the enabled_apis
attribute, and put out a warning if it's included.  The api-metadata
receipe should be used to control this service.  By default the
api-metadata recipe is included in the os-compute-api role which
is included in the os-compute-single-controller role. So, in most cases
it's already included in the run list.

Change-Id: Ief1a98cf2ab072a8aa892c65c5fc07d12a02722c
Closes-Bug: #1415220