Remove docker-setup receipe
The docker-setup receive has been unmaintained and untested for some
time now. The nova-docker project is also being removed from
OpenStack [1] and people are advised to look at deploying Zun as
a replacement.

This also avoids having to rewrite this receipe in order to remove the
use of the deprecated python cookbook.

[1] https://review.openstack.org/430232

Change-Id: Ia8c9db272c6a0353130fc93d446b385fd72ac250
Partial-Bug: 1558538